Fix small bug in `scripts/setup-examples.sh`
Previously, when multiple examples were included, we were `cd`-ing back to the
`$ROOT_DIR` instead of `$ROOT_DIR/examples`. This behavior broke-down when
multiple examples existed.

This was not an issue with this repository because we have only included one
example to this point, but after implementing a fix in another repository, this
fix was applied here as well.